requirements.txt
Django==5.0.7
requests==2.32.3
rasa[spacy]==3.6.20
django-cors-headers==4.4.0
singleton-decorator==1.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1